refactor: 역할을 분리하기 위해서 ModalComponent로 나눈다
